fix: populate gencost_orig and move conversion functions to separate module
